+++ test/all.test.js
@@ -8,8 +8,7 @@ var chai = require('chai'),
 	expect = chai.expect,
 	promised = require('chai-as-promised'),
 	Support = require(__dirname + '/support'),
-	Sequelize = Support.Sequelize,
-	Promise = Sequelize.Promise;
+	Sequelize = Support.Sequelize;
 
 // init
 chai.use(promised);
@@ -17,5 +16,11 @@ chai.config.includeStack = true;
 
 // tests
 
+/* jshint expr: true */
+/* global describe, it */
+
 describe(Support.getTestDialectTeaser('Tests'), function () {
+	it('Tests', function() {
+		expect(Sequelize).to.be.ok;
+	});
 });