+++ changelog.md
@@ -138,5 +138,6 @@ Accidental commit. No changes made.
 
 ## Next
 
+* Run jshint on tests
 * Disable Travis dependency cache
 * Update README badges to use shields.io